backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.29.0
httpx==0.28.1
websockets==15.0.1
qrcode[pil]==7.4.2
PyJWT>=2.10.1
redis==4.6.0
fastapi-limiter==0.1.5
python-json-logger>=2.0.7
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-mock>=3.12.0
passlib>=1.7.4
bcrypt>=4.1.2
deepface>=0.0.92
tensorflow>=2.15.0,<2.17.0
tf-keras>=2.15.0

# Optional: only needed if using Celery for background tasks
celery==5.3.4
# redis==5.0.1
